pyproject.toml
[project]
name = "mnist-ffn-numpy"
version = "0.1.0"
description = "A FFN from scratch with numpy to classify MNIST data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "matplotlib>=3.10.6",
    "mlflow>=3.3.2",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pytest>=8.4.1",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
]
